# OBPIH-6742 Performance: Virtualize table on receiving workflow