@EWaterman 2024-11-12
thanks for the in-depth explanation! It definitely makes sense to want to reload rows one at a time instead of the whole table at once. Great job!